% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{cutLineDownstream}
\alias{cutLineDownstream}
\title{Cut off the downdtream segment of a river segment at a point}
\usage{
cutLineDownstream(line, pt)
}
\arguments{
\item{line}{a spatial line.}

\item{pt}{a spatial point, previously snapped to \code{line}.}
}
\value{
A spatial lines object.
}
\description{
Cut off the downdtream segment of a river segment at a point
}
\examples{
bng <- CRS("+init=epsg:27700")
cc <- cbind(c(0,0), c(0,1))
line <- SpatialLines(list(Lines(list(Line(cc)), ID = "A")), bng)
pt <- SpatialPoints(cbind(0, .5), bng)
down_line <- cutLineDownstream(line, pt)

plot(line)
points(cc)
points(pt, col = "green")
lines(down_line, col = "red")

cc <- cbind(-10:10, y(-10:10))
line <- SpatialLines(list(Lines(list(Line(cc)), ID = "A")), bng)
pt <- SpatialPoints(cbind(-8.020474, 6.493534), bng)
pt <- snapPointsToLines(pt, line)
down_line <- cutLineDownstream(line, pt)

plot(down_line, col = "red", lwd = 2, xlim = c(-10, -2), ylim = c(5, 10))
lines(line)
points(cc)
points(pt, col = "green")
}

