% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutLineUpstream.R
\name{cutLineUpstream}
\alias{cutLineUpstream}
\title{Cut off the upstream segment of a river segment at a point}
\usage{
cutLineUpstream(line, pt)
}
\arguments{
\item{line}{a spatial line.}

\item{pt}{a spatial point, previously snapped to \code{line}.}
}
\value{
A spatial lines object.
}
\description{
Cut off the upstream segment of a river segment at a point
}
\examples{
bng <- CRS("+init=epsg:27700")
cc <- cbind(c(0,0), c(0,1))
line <- SpatialLines(list(Lines(list(Line(cc)), ID = "A")), bng)
pt <- SpatialPoints(cbind(0, .5), bng)
up_line <- cutLineUpstream(line, pt)

plot(line)
points(cc)
points(pt, col = "green")
lines(up_line, col = "red")
y <- function(x) 0.1*x^2

cc <- cbind(-10:10, y(-10:10))
line <- SpatialLines(list(Lines(list(Line(cc)), ID = "A")), bng)
pt <- SpatialPoints(cbind(-8.020474, 6.493534), bng)
pt <- snapPointsToLines(pt, line)
up_line <- cutLineUpstream(line, pt)

plot(up_line, col = "red", lwd = 2, xlim = c(-10, -2), ylim = c(5, 10))
lines(line)
points(cc)
points(pt, col = "green")
}
