% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_functions.R
\name{getPointUpstream}
\alias{getPointUpstream}
\title{Find a point a given distance up stream on a river segment}
\usage{
getPointUpstream(sldf, distance = 50)
}
\arguments{
\item{sldf}{a spatial line.}

\item{distance}{the distance to move upstream.}
}
\value{
A spatial points object.
}
\description{
Find a point a given distance up stream on a river segment
}

